+++ app/gitlab/loading.tsx
@@ -1,4 +1,4 @@
-import LoadingConnectedService from "../../utils/services/loading";
+import LoadingConnectedService from "../../components/services/loading";
 
 export default function loadingConnetedService() {
   return <LoadingConnectedService />;
